convert to controller
