+++ src/lerobot/robots/unitree_g1/controllers/__init__.py
@@ -1,0 +1,8 @@
+"""Unitree G1 locomotion controllers (Groot, Holosoma, SONIC)."""
+
+__all__ = [
+    "GrootLocomotionController",
+    "HolosomaLocomotionController",
+    "SonicWholeBodyController",
+    "SonicRuntime",
+]
